app/requirements.txt
aiohttp==3.6.2
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.4
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.0
cssselect==1.1.0
decorator==4.4.2
dicttoxml==1.7.4
Django==3.1
django-debug-toolbar==2.2
django-redis==4.12.1
django-reverse-admin==2.9.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.16
drf-yasg==1.17.1
future==0.18.2
geocoder==1.38.1
geoip2==4.0.2
idna==2.10
inflection==0.5.0
IP2Location==8.4.1
isort==5.3.0
itypes==1.2.0
jeepney==0.4.3
Jinja2==2.11.2
keyring==21.3.0
lazy-object-proxy==1.5.1
lxml==4.5.2
MarkupSafe==1.1.1
maxminddb==2.0.2
mccabe==0.6.1
multidict==4.7.6
numpy==1.19.1
packaging==20.4
pandas==1.1.0
Pillow==7.2.0
pip-review==1.1.0
pkg-resources==0.0.0
pkginfo==1.5.0.1
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.3
pyparsing==2.4.7
pyquery==1.4.1
python-dateutil==2.8.1
pytz==2020.1
ratelim==0.1.6
readme-renderer==26.0
redis==3.5.3
requests==2.24.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
SecretStorage==3.1.2
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
tqdm==4.48.2
twine==3.2.0
typed-ast==1.4.1
typing==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
webencodings==0.5.1
wrapt==1.12.1
yarl==1.5.1
